// src/types/consignee.d.ts
export interface IConsignee {
    consignee_id?: number;
    order_id?: string;
    name?: string;
    ename?: string;
    company?: string;
    country_code?: string;
    province_name?: string;
    city_name?: string;
    area_name?: string;
    address?: string;
    post_code?: string;
    mobile?: string;
    phone?: string;
    email?: string;
    certificate_type?: string;
    certificate_number?: string;
}